+++ src/IRepository.ts
@@ -1,10 +1,12 @@
 import type { Entity } from './Entity';
 import type { IReadonlyRepository } from './IReadonlyRepository';
 import type {
-  CreateUpdateDeleteOptions, //
+  CreateUpdateOptions, //
+  DeleteOptions,
   DestroyResult,
   DoNotReturnRecords,
   WhereQuery,
+  ReturnSelect,
 } from './query';
 
 export interface IRepository<T extends Entity> extends IReadonlyRepository<T> {
@@ -12,17 +14,17 @@ export interface IRepository<T extends Entity> extends IReadonlyRepository<T> {
    * Creates a objects using the specified values
    * @param {object} values - Values to insert as multiple new objects.
    * @param {{returnRecords: false}} options
-   * @returns {boolean}
+   * @returns {object}
    */
-  create(values: Partial<T>, options?: CreateUpdateDeleteOptions): Promise<T>;
+  create(values: Partial<T>, options?: ReturnSelect): Promise<T>;
 
   /**
    * Creates a objects using the specified values
    * @param {object|object[]} values - Values to insert as multiple new objects.
    * @param {{returnRecords: false}} options
-   * @returns {boolean}
+   * @returns {void}
    */
-  create(values: Partial<T> | Partial<T>[], options: DoNotReturnRecords): Promise<boolean>;
+  create(values: Partial<T> | Partial<T>[], options: DoNotReturnRecords): Promise<void>;
 
   /**
    * Creates a objects using the specified values
@@ -30,9 +32,9 @@ export interface IRepository<T extends Entity> extends IReadonlyRepository<T> {
    * @param {object} [options]
    * @param {boolean} [options.returnRecords=true] - Determines if inserted records should be returned
    * @param {string[]} [options.returnSelect] - Array of model property names to return from the query.
-   * @returns {boolean}
+   * @returns {object[]}
    */
-  create(values: Partial<T>[], options?: CreateUpdateDeleteOptions): Promise<T[]>;
+  create(values: Partial<T>[], options?: ReturnSelect): Promise<T[]>;
 
   /**
    * Creates an object using the specified values
@@ -40,18 +42,18 @@ export interface IRepository<T extends Entity> extends IReadonlyRepository<T> {
    * @param {object} [options]
    * @param {boolean} [options.returnRecords=true] - Determines if inserted records should be returned
    * @param {string[]} [options.returnSelect] - Array of model property names to return from the query.
-   * @returns {object} Return value from the db
+   * @returns {object|object[]|void} Return value from the db
    */
-  create(values: Partial<T> | Partial<T>[], options?: CreateUpdateDeleteOptions): Promise<T | T[] | boolean>;
+  create(values: Partial<T> | Partial<T>[], options?: CreateUpdateOptions): Promise<T | T[] | void>;
 
   /**
    * Updates object(s) matching the where query, with the specified values
    * @param {object} where - Object representing the where query
    * @param {object} values - Values to update
    * @param {{returnRecords: false}} options
-   * @returns {boolean}
+   * @returns {void}
    */
-  update(where: WhereQuery, values: Partial<T>, options: DoNotReturnRecords): Promise<boolean>;
+  update(where: WhereQuery, values: Partial<T>, options: DoNotReturnRecords): Promise<void>;
 
   /**
    * Updates object(s) matching the where query, with the specified values
@@ -60,9 +62,9 @@ export interface IRepository<T extends Entity> extends IReadonlyRepository<T> {
    * @param {object} [options] - Values to update
    * @param {boolean} [options.returnRecords=true] - Determines if inserted records should be returned
    * @param {string[]} [options.returnSelect] - Array of model property names to return from the query.
-   * @returns {boolean}
+   * @returns {object[]}
    */
-  update(where: WhereQuery, values: Partial<T>, options?: CreateUpdateDeleteOptions): Promise<T[]>;
+  update(where: WhereQuery, values: Partial<T>, options?: ReturnSelect): Promise<T[]>;
 
   /**
    * Updates object(s) matching the where query, with the specified values
@@ -71,34 +73,34 @@ export interface IRepository<T extends Entity> extends IReadonlyRepository<T> {
    * @param {object} [options]
    * @param {boolean} [options.returnRecords=true] - Determines if inserted records should be returned
    * @param {string[]} [options.returnSelect] - Array of model property names to return from the query.
-   * @returns {object[]} Return values from the db or `true` if returnRecords=false
+   * @returns {object[]|void} Return values from the db or `true` if returnRecords=false
    */
-  update(where: WhereQuery, values: Partial<T>, options?: CreateUpdateDeleteOptions): Promise<T[] | boolean>;
+  update(where: WhereQuery, values: Partial<T>, options?: CreateUpdateOptions): Promise<T[] | void>;
+
+  /**
+   * Destroys object(s) matching the where query
+   * @param {object} [where] - Object representing the where query
+   * @returns {void}
+   */
+  destroy(where?: WhereQuery): DestroyResult<T, void>;
 
   /**
    * Destroys object(s) matching the where query
    * @param {object} where - Object representing the where query
-   * @param {{returnRecords: false}} options
-   * @returns {boolean}
+   * @param {object} options - Determines if inserted records should be returned
+   * @param {boolean} [options.returnRecords] - Determines if inserted records should be returned
+   * @param {string[]} [options.returnSelect] - Array of model property names to return from the query.
+   * @returns {object[]}
    */
-  destroy(where: WhereQuery, options: DoNotReturnRecords): DestroyResult<T, boolean>;
-
-  /**
-   * Destroys object(s) matching the where query
-   * @param {object} where - Object representing the where query
-   * @param {object} [options] - Determines if inserted records should be returned
-   * @param {boolean} [options.returnRecords=true] - Determines if inserted records should be returned
-   * @returns {boolean}
-   */
-  destroy(where?: WhereQuery, options?: CreateUpdateDeleteOptions): DestroyResult<T, T[]>;
+  destroy(where: WhereQuery, options: DeleteOptions): DestroyResult<T, T[]>;
 
   /**
    * Destroys object(s) matching the where query
    * @param {object} where - Object representing the where query
    * @param {object} [options]
-   * @param {boolean} [options.returnRecords=true] - Determines if inserted records should be returned
+   * @param {boolean} [options.returnRecords=false] - Determines if inserted records should be returned
    * @param {string[]} [options.returnSelect] - Array of model property names to return from the query.
-   * @returns {object[]|boolean} Records affected or `true` if returnRecords=false
+   * @returns {object[]|void} `void` or records affected if returnRecords=true
    */
-  destroy(where: WhereQuery, options?: CreateUpdateDeleteOptions): DestroyResult<T, T[] | boolean>;
+  destroy<TOptions extends DeleteOptions = DeleteOptions>(where: WhereQuery, options?: TOptions): DestroyResult<T, TOptions extends DeleteOptions ? void : T[]>;
 }