+++ src/query/CreateUpdateOptions.ts
@@ -1,4 +1,4 @@
 import type { DoNotReturnRecords } from './DoNotReturnRecords';
 import type { ReturnSelect } from './ReturnSelect';
 
-export type CreateUpdateDeleteOptions = DoNotReturnRecords | ReturnSelect;
+export type CreateUpdateOptions = DoNotReturnRecords | ReturnSelect;