+++ src/query/DeleteOptions.ts
@@ -1,0 +1,11 @@
+interface ReturnSelect {
+  returnSelect: string[];
+  returnRecords?: true;
+}
+
+interface ReturnRecords {
+  returnRecords: true;
+  returnSelect?: string[];
+}
+
+export type DeleteOptions = ReturnSelect | ReturnRecords;